Refactor Connectors
This also fixes some breaking changes in the async-io landscape. An
async_dup::Arc isn't entirely sufficient any more. So we'll use their
Mutex and then wrap it in the Arc to split the stream.

An idea could be to use future-utils bilock -- but it brings in
pin-project. Their implementation is rather simple (basically an
arc-swap). So we could write our own if we want to get rid of that weird
HRTB that async-dup requires