reimplement tags builder
this introduces a 'test' module which'll be a place for user-facing
useful utilities.

the tags builder can both incrementally build tags and merge existing
tags.

this requires escaping and unescaping strings. so two helper functions
are provided.

in the near future get() will unescape by default -- but currently
get_unescaped() provides that role.